+++ packages/notices/CHANGELOG.md
@@ -1,8 +1,13 @@
+## 2.0.0 (Unreleased)
+
+### Breaking Change
+
+- Revert: The `speak` option added in v1.1.0 has been removed.
+
 ## 1.1.0 (2018-11-20)
 
 ### New Feature
 
-- The `createNotice` can now optionally accept a WPNotice object as the sole argument.
 - New option `speak` enables control as to whether the notice content is announced to screen readers (defaults to `true`)
 
 ### Bug Fixes